==> notebooks/1_Data_Cleaning.ipynb <==
import pandas as pd
import pickle

# Data Cleaning

# Import CSV – 4.1M rows of request from 2008 - 2020
df = pd.read_csv('../data/sf_311_raw.csv')

# Sort dataframe by CaseID (newest first)
df.sort_values(by='CaseID', ascending=False, inplace=True)

# Reset the index
df.reset_index(drop=True, inplace=True)

df.shape

# Drop all computed columns (unnecessary)
df = df.iloc[:, 0:20]
# df = df.loc[:, ~df.columns.str.contains('^:@')]

# Drop 'Updated' column
df.drop(columns='Updated', inplace=True)

print(df.shape)
df.head()

# Strip column names and format
df.columns = df.columns.str.strip().str.lower().str.replace(
    ' ', '_').str.replace('(', '').str.replace(')', '')

# Rename 'caseid' column
df.rename({'caseid': 'case_id'}, axis='columns', inplace=True)

df.info()

# Convert time columns to datetime
df['opened'] = pd.to_datetime(df['opened'])
df['closed'] = pd.to_datetime(df['closed'])
# df.iloc[:, 1:4] = pd.to_datetime(df.iloc[:, 1:4].stack()).unstack()

df.shape

# Convert categorical columns to 'category' type
df['supervisor_district'] = pd.to_numeric(df['supervisor_district'], downcast='integer').astype('category')

cols = df[['category', 'neighborhood', 'police_district', 'source']]
for col in cols:
    df[col] = df[col].astype('category')
    
df.info()

# Pickle dataframe
df.to_pickle('../data/df_columns_to_datetime.pkl')

# Load dataframe
df = pd.read_pickle('../data/df_columns_to_datetime.pkl')

# Remove Lat/Long Outliers
df = df.loc[(df['latitude']<37.84) & (df['latitude']>37.70) & (df['longitude']<-122.36) & (df['longitude']>-122.51)]

# Create 'has_media' column 
df['has_media'] = df['media_url'].notnull().astype('int8').astype('category')

# Drop 'media_url' as we now have 'has_media'
df.drop(columns=['media_url'], inplace=True)

# Drop all rows with any missing value
df.dropna(how='any', inplace=True)

# Target Variable

SF311 appears to have a problem with requests that are not resolved since there isn't enough information or worse, the issue is not found once a team is on site. 

# Defining "wasteful" requests
duplicates = ['dup', 'dupe', 'duplicate', 'same', 'already', 'abated']

invalid = ['cancelled', 'case is invalid', 'no action required',
           'insufficient information provided', 'not accepted']

waste = ['not found', 'nothing found', 'none', 'gone on arrival', 'unable to locate', 'no longer',
         'no condition', 'no action', 'no response', 'no construction']

# Create target variable – wasteful requests
def create_target(df):
    """Return 'target' column"""
    regex_str = '''dup|same|already|abated|invalid|cancelled|insufficient|unable|gone|no\b|none\b|not\b|nothing'''
    df.insert(loc=1, column='target',
               value=df['status_notes'].str.lower().str.contains(regex_str, regex=True).astype('int8'))
    return df
    
create_target(df)

df.head()

# Reduce scope
df = df.loc[(df['opened'] >= '2018-01-01') & (df['closed'] <= '2018-12-31')]

df.shape

# Target variable
target_count = df['target'].value_counts()

# Print class balance
print(f'Class 0: {target_count[0]}')
print(f'Class 1: {target_count[1]}')
print(f'Proportion: {round(target_count[0] / target_count[1], 2)} : 1')
print(f'Percentage of Majority Class: {round(target_count[0] / sum(target_count), 3)*100}')

# Pickle
df.to_pickle('../data/01_df_clean_2018.pkl')

# SODA API (Alternative to CSV)
If we needed data to be updated frequently, the SODA API would be preferred to a CSV download.

# from sodapy import Socrata
# # Unauthenticated client only works with public data sets. Note 'None'
# # in place of application token, and no username or password:
# client = Socrata('data.sfgov.org', None)

# # Authenticate the client
# domain = 'data.sfgov.org'
# app_token = '****************'
# username  = 'corraljrmiguel@gmail.com'
# password  = '****************'

# client = Socrata(domain,
#                  app_token,
#                  username=username,
#                  password=password)

# # Returns as JSON from API
# # converted to Python list of dictionaries by sodapy.
# results = client.get('vw6y-z8j6', limit=10000)
# # results = client.get('vw6y-z8j6', where='service_request_id=12167455')

# # Convert to pandas dataframe
# df = pd.DataFrame.from_records(results)
