==> notebooks/2_Feature_Engineering.ipynb <==
import pandas as pd
import numpy as np
import re

from matplotlib.pyplot import figure
import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline

# Load dataframe
df = pd.read_pickle('../data/01_df_clean_2018.pkl')

df.head()

# Latitude and Longitude

# Rounding the lat/long
df['latitude'] = round(df['latitude'], 3)
df['longitude'] = round(df['longitude'], 3)
df.shape

## Future Work (cluster coordinates)

# df['point'] = df['latitude'].apply(str) + ', ' + df['longitude'].apply(str)
# # df['point'] = df['point'].astype('category')
# df.head()
# df.point.nunique()

# from sklearn.cluster import DBSCAN
# from sklearn import metrics

# test = df.head(10000)

# coords = test.loc[:, ['longitude', 'latitude']].to_numpy()

# # kms_per_radian = 500
# # epsilon = 1.5 / kms_per_radian
# db = DBSCAN(eps=.000019, min_samples=5, algorithm='ball_tree', metric='haversine', n_jobs=-1).fit(np.radians(coords)) # eps=.000019 gives 38 clusters for 1000
# cluster_labels = db.labels_
# num_clusters = len(set(cluster_labels))
# clusters = pd.Series([coords[cluster_labels == n] for n in range(num_clusters)])

# print('Number of clusters: {}'.format(num_clusters))

# test = pd.DataFrame(cluster_labels)
# test
# test[0].unique()

# Cyclical Features
We need to represent the cyclical features as (x,y) coordinates on a circle.  
http://blog.davidkaleko.com/feature-engineering-cyclical-features.html

# Year
df['opened_year'] = df['opened'].dt.year.astype('category')

# Month
month = df['opened'].dt.month
df['opened_month_sin'] = np.sin(month*(2.*np.pi/12))
df['opened_month_cos'] = np.cos(month*(2.*np.pi/12))

# Week
week = df['opened'].dt.week
df['opened_week_sin'] = np.sin(week*(2.*np.pi/52))
df['opened_week_cos'] = np.cos(week*(2.*np.pi/52))

# Day of week
day = df['opened'].dt.dayofweek
df['opened_day_sin'] = np.sin(day*(2.*np.pi/7))
df['opened_day_cos'] = np.cos(day*(2.*np.pi/7))

# Hour
hour = df['opened'].dt.hour
df['opened_hour_sin'] = np.sin(hour*(2.*np.pi/24))
df['opened_hour_cos'] = np.cos(hour*(2.*np.pi/24))

df.head()

# Resolution Time

# Diff in minutes
df['resolution_time'] = df['closed'] - df['opened']
df['resolution_time'] = (df['resolution_time']/np.timedelta64(1,'m'))
# Remove rows with negative minutes & requests over 7 days
df = df.loc[(df['resolution_time'] > 0)] #& (df['resolution_time'] < 10080)]
df.shape

# Media

# # Create 'has_media' column 
# df['has_media'] = df['media_url'].notnull().astype('int8').astype('category')

# Category Groups

re_street = ['street_cleaning', 'sidewalk cleaning']
re_graffiti = ['graffiti', 'graffiti']
re_parking = ['parking', 'abandoned vehicle|parking enforcement|blocked street or sidewalk']
re_general = ['general', 'public works|sewer issues|tree\b|streetlights|curb|noise']
re_home = ['homelessness', 'encampments|homeless']
re_muni = ['muni_feedback', 'muni']

re_list = [re_street, re_graffiti, re_parking, re_general, re_home, re_muni]

for i in re_list:
    df.insert(loc=(len(df.columns)), column=i[0],
              value=df['category'].str.lower().str.contains(i[1], regex=True).astype('int8'))

df.to_pickle('../data/02_df_pre_model_2018.pkl')

# Appendix

## Category Grouping
* General
    * Blocked Pedestrian Walkway
    * Encampment
    * Noise Issue
    * Park Requests
    * Tree Maintenance
* Cleaner Streets
    * Garbage Containers
    * Graffiti
    * Illegal Postings
    * Street or Sidewalk Cleaning
* Parking & Transportation
    * Abandoned Vehicles
    * Blocked Driveway & Illegal Parking
    * Muni Employee Feedback
    * Muni Service Feedback
* Repair
    * Curb & Sidewalk Issues
    * Damaged Public Property
    * Flooding, Sewer & Water Leak Issues
    * Parking & Traffic Sign Repair
    * Pothole & Street Issues
    * Streetlight Repair

# Distribution of Resolution Time
plt.hist(df['resolution_time'])

# Top categories by frequency of wasteful requests
df.loc[df['target']==1, 'category'].value_counts()[:10]
