==> notebooks/3_Feature_Selection.ipynb <==
import pandas as pd
import numpy as np

# Tools
from collections import Counter
import pickle

# Preprocessing & Selections
from sklearn.preprocessing import LabelEncoder, OneHotEncoder
from sklearn.preprocessing import MinMaxScaler
from sklearn.feature_selection import SelectKBest, chi2, f_classif
from sklearn.model_selection import train_test_split

# Sampling
from imblearn.over_sampling import SMOTE
from imblearn.under_sampling import NearMiss
from imblearn.under_sampling import RandomUnderSampler

# Load dataframe
df = pd.read_pickle('../data/02_df_pre_model_2018.pkl')

# # Convert to Dask dataframe
# df = dd.from_pandas(df_pd, npartitions=16)
df.head()

# Train and test splitting

# Columns to exclude
exclude_cols = [
    'target', # Target variable
    'case_id',
    'opened', # Feature Eng
    'closed', # Feature Eng
    'status',
    'status_notes', # Needs NLP
    'request_details', # Needs NLP
    'address', # Needs NLP
#     'street',
    'point',

    # New items
    'responsible_agency',
    'category', # Need to choose 'category' or 'request_type' NOT BOTH
#     'request_type', # Needs NLP
    'opened_year',
#     'opened_month_sin',
#     'opened_month_cos',
#     'opened_week_sin',
#     'opened_week_cos',
#     'opened_day_sin',
#     'opened_day_cos',
#     'opened_hour_sin',
#     'opened_hour_cos',
    'police_district',
    'supervisor_district',
#     'latitude',
    'longitude',
]

# Predictor variables
X = df.drop(columns=exclude_cols, axis=0, inplace=False)

# Get dummies for categorical variables
X = pd.get_dummies(X, drop_first=True)

# Target variable
y = df['target']

# Split train and test
X_train, X_test, y_train, y_test = train_test_split(X, y,
                                                    test_size=0.2, 
                                                    random_state=2020, 
                                                    stratify=y,  # Stratify to keep same class ratios
                                                    shuffle=True # Shuffle data since it's ordered chronologically
                                                   )
X_train.head()

scaler = MinMaxScaler()
scaler.fit(X_train)
X_train_scaled = scaler.transform(X_train)
X_test_scaled = scaler.transform(X_test)

#Medium
# scaler = StandardScaler()
# scaler.fit(X_train)
# X_train = scaler.transform(X_train)
# X_test = scaler.transform(X_test)



# StackOverflow
# scaler = MinMaxScaler()
# X_train_scaled = scaler.fit_transform(X_train)

# model = SVC()
# model.fit(X_train_scaled, y_train)

# X_test_scaled = scaler.transform(X_test)
# y_pred = model.predict(X_test_scaled)

# # Pickle for later use
# with open('../data/03_X.pkl', 'wb') as f:
#     pickle.dump(X, f)
#     f.close()

# X_train.to_pickle('../data/X_train.pkl')
# X_test.to_pickle('../data/X_test.pkl')
# y_train.to_pickle('../data/y_train.pkl')
# y_test.to_pickle('../data/y_test.pkl')

# Feature Selection

def select_features(X_train, y_train, X_test):
    '''Returns X_train, X_test, and feature selection function'''
    fs = SelectKBest(score_func=chi2, k='all')
    fs.fit(X_train, y_train)
#     X_train_fs = fs.transform(X_train)
#     X_test_fs = fs.transform(X_test)
#     return X_train_fs, X_test_fs, fs
    return fs

# Feature selection
# X_train_fs, X_test_fs, fs = select_features(X_train, y_train, X_test)
fs = select_features(X_train_scaled, y_train, X_test_scaled)

# # Feature scores
# features_df = pd.DataFrame(data=[X_train.columns, fs.scores_.astype(int)]).transpose()
# features_df.rename(columns={0: 'Feature', 1: 'ANOVA F-Value'}, inplace=True)
# features_df.sort_values(by='ANOVA F-Value', ascending=False, inplace=True)
# features_df.reset_index(drop=True, inplace=True)
# features_df

# Feature scores
features_df = pd.DataFrame(data=[X_train.columns, fs.scores_.astype(int)]).transpose()
features_df.rename(columns={0: 'Feature', 1: 'Chi2'}, inplace=True)
features_df.sort_values(by='Chi2', ascending=False, inplace=True)
features_df.reset_index(drop=True, inplace=True)
features_df

# Select features above threshold
threshold = 50
# best_features_df = features_df[(features_df['ANOVA F-Value'] > threshold)]
best_features_df = features_df[(features_df['Chi2'] > threshold)]
best_features_df

# best_features_df.to_pickle('../data/best_features_df.pkl')
# best_features_df = pd.read_pickle('../data/best_features_df.pkl')

# # Filter X_train & X_test with selected features
# X_train = X_train.filter(items=best_features_df['Feature'])
# X_test  = X_test.filter(items=best_features_df['Feature'])

# # Clean column names
# X_train.columns = X_train.columns.str.strip().str.lower().str.replace(
#     ' ', '_').str.replace('(', '').str.replace(')', '')

# X_test.columns = X_test.columns.str.strip().str.lower().str.replace(
#     ' ', '_').str.replace('(', '').str.replace(')', '')

# Filter X_train & X_test with selected features
X_train_scaled = pd.DataFrame(X_train_scaled, columns=X_train.columns).filter(items=best_features_df['Feature'])
X_test_scaled  = pd.DataFrame(X_test_scaled, columns=X_train.columns).filter(items=best_features_df['Feature'])

# Clean column names
X_train_scaled.columns = X_train_scaled.columns.str.strip().str.lower().str.replace(
    ' ', '_').str.replace('(', '').str.replace(')', '')

X_test_scaled.columns = X_test_scaled.columns.str.strip().str.lower().str.replace(
    ' ', '_').str.replace('(', '').str.replace(')', '')

print('df\t', df.shape)
print('X_train\t', X_train_scaled.shape)
print('X_test\t', X_test_scaled.shape)
print('y_train\t', y_train.shape)
print('y_test\t', y_test.shape)

# Class Balancing

# Target variable
target_count = df['target'].value_counts()

# Print class balance
print(f'Class 0: {target_count[0]}')
print(f'Class 1: {target_count[1]}')
print(f'Proportion: {round(target_count[0] / target_count[1], 2)} : 1')
print(f'Percentage of Majority Class: {round(target_count[0] / sum(target_count), 3)*100}')

## Oversampling

# # Define the oversampling method – SMOTE
# smote = SMOTE(random_state=2020)
# X_train_smote, y_train_smote = smote.fit_sample(X_train, y_train)

# # Summarize the new class distribution
# Counter(y_train_smote)

## Undersampling

# Define the undersampling method – RandomUnderSampler
rndm_under = RandomUnderSampler(random_state=2020)

# Transform the dataset
# X_train_under, y_train_under = rndm_under.fit_sample(X_train, y_train)
X_train_under, y_train_under = rndm_under.fit_sample(X_train_scaled, y_train)

# New class distribution
Counter(y_train_under)

# Pickle dataframes
df.to_pickle('../data/df.pkl')
X_train_under.to_pickle('../data/03_X_train_under.pkl')
X_test_scaled.to_pickle('../data/03_X_test.pkl')
y_train_under.to_pickle('../data/03_y_train_under.pkl')
y_test.to_pickle('../data/03_y_test.pkl')

# # Transform to Dask dataframes
# X_train_under = dd.from_pandas(X_train_under, npartitions=16)
# X_test        = dd.from_pandas(X_test, npartitions=16)
# y_train_under = dd.from_pandas(y_train_under, npartitions=16)
# y_test        = dd.from_pandas(y_test, npartitions=16)

# Appendix

# Dask
# cat /proc/cpuinfo
# from dask.distributed import Client, progress
# from sklearn.externals.joblib import parallel_backend

# client = Client(processes=False)
# # client = Client(processes=False, n_workers=4, threads_per_worker=8)
# client
# # client.close()

# # Define the undersampling method – NearMiss
# # Selects the closest examples from the majority class for each minority class.
# undersample = NearMiss(version=3, n_neighbors_ver3=3)

# # Transform the dataset
# X_train_under, y_train_under = undersample.fit_resample(X_train, y_train)

# # Summarize the new class distribution
# Counter(y_train_under)
